fix(sync) fixed the problems from resolving merge conflicts
