+++ screens/NewWorkspaceScreen/NewWorkspaceScreen.js
@@ -58,7 +58,7 @@ const NewWorkspaceScreen = (props) => {
       </View>
       <View style={styles.stepContent}>
         <Text style={styles.subTitle}>
-          Paso 2 
+          Paso 2
         </Text>
         <Text style={styles.row}>
         Copia el código mostrado al final de la autorización e ingrésalo en el siguiente campo.
@@ -76,16 +76,15 @@ const NewWorkspaceScreen = (props) => {
       </View>
       <View style={styles.stepContent}>
         <Text style={styles.subTitle}>
-          Paso 3 
+          Paso 3
         </Text>
-        <Text style={styles.row}> 
+        <Text style={styles.row}>
         Presiona listo.
         </Text>
       </View>
-      <View style={styles.bottomContent}>
-        <View style={styles.buttonContainer}>
-          <TouchableOpacity style={styles.button} onPress={pressHandler}>
-            <Text style={styles.textButton}>
+      <View style={styles.buttonContainer}>
+        <TouchableOpacity style={styles.button} onPress={pressHandler}>
+          <Text style={styles.textButton}>
 
             Listo
           </Text>
